% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{alt_dnorm}
\alias{alt_dnorm}
\title{Alternative Functions}
\usage{
alt_dnorm(x, mean = 0, sd = 1)
}
\arguments{
\item{x, mean, sd}{See \code{norm}.}
}
\value{
See \code{norm}.
}
\description{
Used for unit tests.
}
