% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{alt_pnorm}
\alias{alt_pnorm}
\title{Alternative Functions}
\usage{
alt_pnorm(q, mean = 0, sd = 1)
}
\arguments{
\item{q, mean, sd}{See \code{norm}.}
}
\value{
See \code{norm}.
}
\description{
Alternative Functions
}
