% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{alt_pt}
\alias{alt_pt}
\title{Alternative Functions}
\usage{
alt_pt(q, df, mean = 0, sd = 1)
}
\arguments{
\item{q, df, mean, sd}{See \code{t}.}
}
\value{
See \code{t}.
}
\description{
Accesses a t-distribution based on the degrees-of-freedom,
 mean, and standard deviation.
}
