% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{alt_qnorm}
\alias{alt_qnorm}
\title{Alternative Functions}
\usage{
alt_qnorm(p, mean = 0, sd = 1)
}
\arguments{
\item{p, mean, sd}{See \code{norm}.}
}
\value{
See \code{norm}.
}
\description{
Used for unit tests.
}
