% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-diverge.R
\name{diverge.kl.sym}
\alias{diverge.kl.sym}
\title{Calculate Divergence between Distributions}
\usage{
diverge.kl.sym(m1, s1, m2, s2)
}
\arguments{
\item{m1, m2}{The means of the distributions.}

\item{s1, s2}{The standard deviations of the distributions.}
}
\value{
The divergence.
}
\description{
Calculate Divergence between Distributions
}
