% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-diverge.R
\name{getDivergences}
\alias{getDivergences}
\title{Get Divergences for a Scheduler Object}
\usage{
getDivergences(scheduler, how = diverge.kl.sym)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{how}{A function. The method used to calculate divergence.}
}
\value{
A matrix of divergences.
}
\description{
Get Divergences for a Scheduler Object
}
