% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{getPermutations}
\alias{getPermutations}
\title{Permute Rewards for a Scheduler Object}
\usage{
getPermutations(scheduler, how = permute.from.all, reference = NULL)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{how}{A function. The method used to permute the rewards.}

\item{reference}{An integer. Passed to \code{how}.}
}
\value{
A \code{scheduler} object.
}
\description{
Permute Rewards for a Scheduler Object
}
