% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{long2wide}
\alias{long2wide}
\title{Make Wide Data from Long Data}
\usage{
long2wide(formula, long)
}
\arguments{
\item{formula}{A formula to guide expansion. The y will be the cell.
The first x will be the row. The last x will be the column.}

\item{long}{A data set in long format.}
}
\value{
A data set in wide format.
}
\description{
Make Wide Data from Long Data
}
