% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-sampler.R
\name{p_greater_than_cutoff}
\alias{p_greater_than_cutoff}
\title{Calculate Probability of Posterior}
\usage{
p_greater_than_cutoff(scheduler, cutoff)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{cutoff}{A numeric. The value against which to compare the posterior.}
}
\value{
A vector of probabilities.
}
\description{
This function estimates the probability that the posterior expected reward
 is greater than a cutoff. It does this by computing the area under the curve
 for a t-score at a given cutoff.
}
