% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{permute.control}
\alias{permute.control}
\title{Permute a List of Rewards}
\usage{
permute.control(rewards, reference = NULL)
}
\arguments{
\item{rewards}{A list of rewards.}

\item{reference}{An integer. The list element with control rewards.}
}
\value{
A list of permuted rewards.
}
\description{
This function samples from the control group only, with replacement.
 The user must select the control group.
}
