% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{permute.from.all}
\alias{permute.from.all}
\title{Permute a List of Rewards}
\usage{
permute.from.all(rewards, reference = NULL)
}
\arguments{
\item{rewards}{A list of rewards.}

\item{reference}{Null argument.}
}
\value{
A list of permuted rewards.
}
\description{
This function samples from any group, without replacement.
}
