% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{permute.from.control.norm}
\alias{permute.from.control.norm}
\title{Permute a List of Rewards}
\usage{
permute.from.control.norm(rewards, reference = NULL)
}
\arguments{
\item{rewards}{A list of rewards.}

\item{reference}{An integer. The list element with control rewards.}
}
\value{
A list of permuted rewards.
}
\description{
This function samples from a normal distribution with the same
 mean and standard deviation as the reference distribution.
}
