% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{permute.null}
\alias{permute.null}
\title{Permute a List of Rewards}
\usage{
permute.null(rewards, reference = NULL)
}
\arguments{
\item{rewards}{A list of rewards.}

\item{reference}{Null argument.}
}
\value{
A list of rewards.
}
\description{
This function does not actually permute the rewards.
 Used for testing purposes only.
}
