% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-stats.R
\name{rewards_as_df}
\alias{rewards_as_df}
\title{Represent Reward List as Data Frame}
\usage{
rewards_as_df(scheduler)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}
}
\value{
A \code{data.frame}
}
\description{
Represent Reward List as Data Frame
}
