% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-run.R
\name{run.benchmark}
\alias{run.benchmark}
\title{Establish the Positive Rate}
\usage{
run.benchmark(
  scheduler,
  simulator,
  N.trials = 10,
  N.allocate = scheduler@N.burn.in,
  alpha = 0.05,
  repititions = 1000,
  how.stats = stats.empiric,
  ...
)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{simulator}{A \code{simulator} object.}

\item{N.trials}{An integer. The total number of times to allocate patients.}

\item{N.allocate}{An integer. The total number of patients to allocate next.
By default, \code{N.allocate} is set equal to \code{N.burn.in}.}

\item{alpha}{The Type I error to control.}

\item{repititions}{An integer. The number of times to repeat the trial.}

\item{how.stats}{A function. The statistical method used to compute a p-value.}

\item{...}{Named arguments passed to \code{how.stats} method.}
}
\value{
A \code{data.frame} of false positive rates.
}
\description{
Establish the Positive Rate
}
