% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-run.R
\name{run.trial}
\alias{run.trial}
\title{Run a Simulated Trial}
\usage{
run.trial(
  scheduler,
  simulator,
  N.trials = 10,
  N.allocate = scheduler@N.burn.in
)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{simulator}{A \code{simulator} object.}

\item{N.trials}{An integer. The total number of times to allocate patients.}

\item{N.allocate}{An integer. The total number of patients to allocate next.
By default, \code{N.allocate} is set equal to \code{N.burn.in}.}
}
\value{
A \code{scheduler} object.
}
\description{
Run a Simulated Trial
}
