% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-simulator.R
\docType{class}
\name{simulator}
\alias{simulator}
\alias{simulator-class}
\alias{show,simulator-method}
\alias{getPool}
\alias{getDraw}
\alias{simulator.start}
\alias{simulator.start.from.norm}
\alias{simulator.start.from.unif}
\alias{simulator.draw}
\title{S4 Simulator Class Object}
\usage{
\S4method{show}{simulator}(object)

getPool(simulator)

getDraw(simulator)

simulator.start(pool)

simulator.start.from.norm(means, variances, N = 1000)

simulator.start.from.unif(mins, maxes, N = 1000)

simulator.draw(simulator, allocation)
}
\arguments{
\item{object, simulator}{A \code{simulator} object.}

\item{pool}{A list of vectors for all rewards available.}

\item{means}{A vector of average rewards, one for each group.}

\item{variances}{A vector of reward variances, one for each group.}

\item{N}{An integer. The size of each group within the pool.}

\item{mins}{A vector of minimums, one for each group.}

\item{maxes}{A vector of maximums, one for each group.}

\item{allocation}{A vector of groups to which to allocate new patients.}
}
\description{
S4 Simulator Class Object
}
\section{Slots}{

\describe{
\item{\code{pool}}{A list of vectors for all rewards available.}

\item{\code{meta}}{A list of tables for all rewards and meta-data available.
This is optional, but may be preferred for complex simulations.}

\item{\code{draw}}{The data most recently sampled from the pool.}
}}

\section{Methods}{

\code{show:} Method to show \code{simulator} object.
}

\section{Getters}{

\code{getPool:} Method to retrieve the data pool from a \code{simulator} object.
 This function returns a list of vectors for all rewards available.


\code{getDraw:} Method to retrieve the previous draw from a \code{simulator} object.
 This function returns the data most recently sampled from the pool.
}

\section{Setters}{

\code{simulator.start:} Method to initialize a \code{simulator} object.
 This function returns a \code{simulator} object.


\code{simulator.start.from.norm:} Method to initialize a \code{simulator} object.
 This function returns a \code{simulator} object.


\code{simulator.start.from.unif:} Method to initialize a \code{simulator} object.
 This function returns a \code{simulator} object.


\code{simulator.draw:} Method to draw data from a \code{simulator} object.
 This function returns a \code{simulator} object.
}

