% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-stats.R
\name{stats.anova}
\alias{stats.anova}
\title{Perform a Statistical Analysis}
\usage{
stats.anova(scheduler)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}
}
\value{
A single p-value.
}
\description{
Perform a Statistical Analysis
}
