% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-stats.R
\name{stats.ttest}
\alias{stats.ttest}
\title{Perform a Statistical Analysis}
\usage{
stats.ttest(scheduler, reference = NULL)
}
\arguments{
\item{scheduler}{A \code{scheduler} object.}

\item{reference}{An integer. The list element with control rewards.}
}
\value{
A vector of adjusted p-values.
}
\description{
Perform a Statistical Analysis
}
