% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-permute.R
\name{structure_as_list}
\alias{structure_as_list}
\title{Structure Vector like a List}
\usage{
structure_as_list(vector, list)
}
\arguments{
\item{vector}{A vector.}

\item{list}{A list.}
}
\value{
A list.
}
\description{
This function structures a vector to look like a list
 based on a reference list.
}
