% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-scheduler-class.R, R/2-scheduler-methods.R,
%   R/3-scheduler-viz.R, R/4-sampler.R
\docType{class}
\name{scheduler}
\alias{scheduler}
\alias{scheduler-class}
\alias{show,scheduler-method}
\alias{getAllocation}
\alias{getHistory}
\alias{scheduler.start}
\alias{scheduler.update}
\alias{plotPrior}
\alias{plotPosterior}
\alias{plotHistory}
\alias{plotAllocation}
\alias{thompson}
\title{S4 Scheduler Class Object}
\usage{
\S4method{show}{scheduler}(object)

getAllocation(object)

getHistory(object)

scheduler.start(prior.mean, prior.var, N.burn.in, sampler = "thompson")

scheduler.update(scheduler, data.ingest, N.allocate)

plotPrior(scheduler)

plotPosterior(scheduler)

plotHistory(scheduler)

plotAllocation(scheduler)

thompson(scheduler)
}
\arguments{
\item{object, scheduler}{A \code{scheduler} object.}

\item{prior.mean}{A vector of prior means. One mean for each arm}

\item{prior.var}{A vector of prior variances. One variance for each arm.}

\item{N.burn.in}{An integer. The number patients to allocate each arm
during the initial 'burn-in' phase. Ideally, around 20-30 per arm.}

\item{sampler}{A string. The name of the sampling method used to allocate patients.}

\item{data.ingest}{A list of rewards. Must match the structure of \code{scheduler@ingest}.}

\item{N.allocate}{An integer. The total number of patients to allocate next.}
}
\description{
S4 Scheduler Class Object
}
\section{Slots}{

\describe{
\item{\code{prior.mean,prior.var,prior.prec}}{The prior statistics for each arm.}

\item{\code{N.burn.in}}{An integer. The number of patients in each arm at first time step.}

\item{\code{K.arms}}{An integer. The number of experimental groups (also called 'arms').}

\item{\code{step}}{An integer. The number of time steps so far. Each time step is a
kind of 'mini-trial'. This initializes at 0.}

\item{\code{rewards}}{A list of all rewards observed to date.}

\item{\code{online.count,online.sum,online.mean,online.var,online.prec}}{The empiric statistics observed to date.}

\item{\code{post.mean,post.var,post.prec}}{The posterior statistics for each arm.}

\item{\code{sampler}}{A string. The name of the sampling method used to allocate patients.}

\item{\code{allocation}}{A vector of groups to which to allocate new patients.}

\item{\code{ingest}}{A list describing the structure of the data expected
by the next \code{scheduler.update} call. This slot is only
used to verify the incoming data.}

\item{\code{allocation}}{The most recent patient allocations. This slot
implies the structure of the \code{@ingest} slot.}

\item{\code{history.post}}{A table of all posterior statistics and the allocation
ratios at each time step in the experiment. At \code{@step = 0},
the table records the prior statistics and burn-in conditions.}

\item{\code{history}}{A list of lists of all rewards. Unlike \code{@rewards},
the rewards here are organized by time step.}
}}

\section{Methods}{

\code{show:} Method to show \code{scheduler} object.
}

\section{Getters}{

\code{getAllocation:} Method to retrieve allocations from a \code{scheduler} object.
 This function returns a vector of groups to which to allocate new patients.


\code{getHistory:} Method to retrieve history from a \code{scheduler} object.
 This function returns a table of all posterior statistics and the allocation
 ratios at each time step in the experiment.
}

\section{Setters}{

\code{scheduler.start:} Method to initiate a \code{scheduler} object.
 This function returns an updated \code{scheduler} object.


\code{scheduler.start:} Method to update a \code{scheduler} object with new data.
 This function returns an updated \code{scheduler} object.
}

\section{Plotters}{

\code{plotPrior:} Method to plot prior distributions.


\code{plotPosterior:} Method to plot the most recent posterior distributions.


\code{plotHistory:} Method to plot all posterior distributions.


\code{plotAllocation:} Method to plot the allocation ratios for each time step.
}

\section{Samplers}{

\code{thompson:} Method to allocate patients by Thompson sampling.
 This function returns an integer corresponding to the group
 to which the patient is randomly allocated.
}

